+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ## [Unreleased]
 ### Added
 - Travis CI integration
+- Boleto PDF layout option
 
 ### Change
 - Test user credentials to environment variable instead of hard-code